+++ docs/content/3.middleware/4.reference/multistore/change-log.md
@@ -1,5 +1,12 @@
 # Change log
 
+## 4.1.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - @vue-storefront/middleware@5.0.0
+
 ## 4.1.0
 
 ### Minor Changes